# apps/web/src/generated/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reframe V7 API",
    "description": "Dual-Run Retrospective Deep Narrative Forensics API with ClickHouse & Google ADK",
    "version": "7.0.0"
  },
  "paths": {
...
    "/api/v1/admin/audience-lab/runs": {
      "post": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Create Audience Simulation Run",
        "description": "Enqueues a durable Audience Simulation analysis run.",
        "operationId": "create_audience_simulation_run_api_v1_admin_audience_lab_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAudienceRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Audience Simulation Run",
        "description": "Fetches status and results of an audience simulation run.",
        "operationId": "get_audience_simulation_run_api_v1_admin_audience_lab_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}/funnel": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Run Funnel",
        "operationId": "get_run_funnel_api_v1_admin_audience_lab_runs__run_id__funnel_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}/bottlenecks": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Run Bottlenecks",
        "operationId": "get_run_bottlenecks_api_v1_admin_audience_lab_runs__run_id__bottlenecks_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}/cohorts": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Run Cohorts",
        "operationId": "get_run_cohorts_api_v1_admin_audience_lab_runs__run_id__cohorts_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}/sensitivity": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Run Sensitivity",
        "operationId": "get_run_sensitivity_api_v1_admin_audience_lab_runs__run_id__sensitivity_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}/load-profile": {
      "get": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Get Run Load Profile",
        "operationId": "get_run_load_profile_api_v1_admin_audience_lab_runs__run_id__load_profile_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audience-lab/runs/{run_id}/seed-content": {
      "post": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Seed Content For Run",
        "operationId": "seed_content_for_run_api_v1_admin_audience_lab_runs__run_id__seed_content_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedContentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audience-lab/demand-scenario": {
      "post": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Calculate Demand Scenario",
        "description": "Calculates directional demand scenario projections strictly from a completed simulation run result.",
        "operationId": "calculate_demand_scenario_api_v1_admin_audience_lab_demand_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectionalDemandInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audience-lab/purge-content": {
      "post": {
        "tags": [
          "Admin Audience Lab"
        ],
        "summary": "Purge Synthetic Content",
        "description": "Safely purges synthetic demo content only, leaving human user content completely untouched.",
        "operationId": "purge_synthetic_content_api_v1_admin_audience_lab_purge_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateAudienceRunRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/SimulationMode",
            "default": "CI"
          },
          "unique_source_personas": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 100.0,
            "title": "Unique Source Personas",
            "default": 10000
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ScenarioVariant"
            },
            "type": "array",
            "title": "Variants"
          },
          "replications": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Replications",
            "default": 1
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "seed_content": {
            "type": "boolean",
            "title": "Seed Content",
            "default": false
          }
        },
        "type": "object",
        "title": "CreateAudienceRunRequest"
      },
...
      "DirectionalDemandInput": {
        "properties": {
          "simulation_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Run Id"
          },
          "forecast_class": {
            "type": "string",
            "title": "Forecast Class",
            "default": "ASSUMPTION_BASED_DIRECTIONAL"
          },
          "monthly_visitors": {
            "type": "integer",
            "title": "Monthly Visitors",
            "default": 50000
          },
          "traffic_source_mix": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Traffic Source Mix"
          },
          "returning_visitor_rate": {
            "type": "number",
            "title": "Returning Visitor Rate",
            "default": 0.25
          },
          "catalog_size": {
            "type": "integer",
            "title": "Catalog Size",
            "default": 1
          },
          "new_article_frequency_monthly": {
            "type": "integer",
            "title": "New Article Frequency Monthly",
            "default": 10
          },
          "new_community_post_frequency_monthly": {
            "type": "integer",
            "title": "New Community Post Frequency Monthly",
            "default": 50
          },
          "campaign_conversion_assumption": {
            "type": "number",
            "title": "Campaign Conversion Assumption",
            "default": 0.05
          },
          "real_user_forecast": {
            "type": "boolean",
            "title": "Real User Forecast",
            "default": false
          }
        },
        "type": "object",
        "title": "DirectionalDemandInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SeedContentRequest": {
        "properties": {
          "posts_count": {
            "type": "integer",
            "maximum": 150.0,
            "minimum": 1.0,
            "title": "Posts Count",
            "default": 20
          },
          "comments_count": {
            "type": "integer",
            "maximum": 800.0,
            "minimum": 1.0,
            "title": "Comments Count",
            "default": 40
          },
          "counterclaims_count": {
            "type": "integer",
            "maximum": 150.0,
            "minimum": 1.0,
            "title": "Counterclaims Count",
            "default": 20
          },
          "reactions_count": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Reactions Count",
            "default": 60
          },
          "enable_magazine": {
            "type": "boolean",
            "title": "Enable Magazine",
            "default": true
          }
        },
        "type": "object",
        "title": "SeedContentRequest"
      },
...
    }
  }
}